--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "classnames": "^2.3.1"
-  }
-}
