+++ project/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@reduxjs/toolkit": "^1.5.0",
     "axios": "^0.21.1",
+    "classnames": "^2.3.1",
     "history": "^5.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
